PP-12356 Move existing tests for PUT /v1/api/service/{serviceId}/[test|live]/3ds-flex-credentials to same location as existing 3ds flex tests
