+++ src/test/java/uk/gov/pay/connector/gatewayaccountcredentials/resource/GatewayAccountCredentialsResourceIT.java
@@ -297,80 +297,6 @@ public class GatewayAccountCredentialsResourceIT {
                     .body("message[0]", is("Operation not supported for payment provider 'epdq'"));
         }
     }
-    
-    @Nested
-    class CreateOrUpdateWorldpay3DSCredentials_byServiceIdAndAccountType {
-        private static final String VALID_ISSUER = "53f0917f101a4428b69d5fb0"; // pragma: allowlist secret`
-        private static final String VALID_ORG_UNIT_ID = "57992a087a0c4849895ab8a2"; // pragma: allowlist secret`
-        private static final String VALID_JWT_MAC_KEY = "4cabd5d2-0133-4e82-b0e5-2024dbeddaa9"; // pragma: allowlist secret`
-        private final Map<String, String> valid3dsFlexCredentialsPayload = Map.of(
-                "issuer", VALID_ISSUER,
-                "organisational_unit_id", VALID_ORG_UNIT_ID,
-                "jwt_mac_key", VALID_JWT_MAC_KEY
-        );
-        @Test
-        void shouldUpdateCredentialsAndReturn200() {
-            String serviceId = "a-valid-service-id";
-            String gatewayAccountId = app.givenSetup()
-                .body(toJson(Map.of(
-                        "payment_provider", "worldpay",
-                        "service_id", serviceId,
-                        "service_name", "a-test-service",
-                        "type", "test"
-                )))
-                .post("/v1/api/accounts")
-                .then().extract().path("gateway_account_id");
-            
-            app.givenSetup()
-                    .body(toJson(valid3dsFlexCredentialsPayload))
-                    .put(format("/v1/api/service/%s/%s/3ds-flex-credentials", serviceId, TEST))
-                    .then()
-                    .statusCode(200);
-
-            Map<String, Object> updatedGatewayAccountCredentials = app.getDatabaseTestHelper().getWorldpay3dsFlexCredentials(Long.valueOf(gatewayAccountId));
-            assertThat(updatedGatewayAccountCredentials, hasEntry("issuer", VALID_ISSUER));
-            assertThat(updatedGatewayAccountCredentials, hasEntry("jwt_mac_key", VALID_JWT_MAC_KEY));
-            assertThat(updatedGatewayAccountCredentials, hasEntry("organisational_unit_id", VALID_ORG_UNIT_ID));
-        }
-        @Test
-        void shouldReturn404_forNonExistentGatewayAccount() {
-            String serviceId = "a-service-id-with-no-gateway-account";
-            
-            app.givenSetup()
-                    .body(toJson(valid3dsFlexCredentialsPayload))
-                    .put(format("/v1/api/service/%s/%s/3ds-flex-credentials", serviceId, TEST))
-                    .then()
-                    .statusCode(404)
-                    .body("message[0]", is(format("Gateway account not found for service ID [%s] and account type [%s]", serviceId, TEST)));
-        }
-        
-        @Test
-        void shouldReturn404_andNotUpdateCredentials_forNonWorldpayGatewayAccount() {
-            String serviceId = "a-valid-service-id";
-            String gatewayAccountId = app.givenSetup()
-                    .body(toJson(Map.of(
-                            "payment_provider", "stripe",
-                            "service_id", serviceId,
-                            "service_name", "a-test-service",
-                            "type", "test"
-                    )))
-                    .post("/v1/api/accounts")
-                    .then().extract().path("gateway_account_id");
-
-            app.givenSetup()
-                    .body(toJson(valid3dsFlexCredentialsPayload))
-                    .put(format("/v1/api/service/%s/%s/3ds-flex-credentials", serviceId, TEST))
-                    .then()
-                    .statusCode(404)
-                    .body("message[0]", is("Not a Worldpay gateway account"));
-
-            Exception exception = assertThrows(IllegalStateException.class, () -> 
-                app.getDatabaseTestHelper().getWorldpay3dsFlexCredentials(Long.valueOf(gatewayAccountId))
-            );
-            assertThat(exception.getMessage(), is("Expected at least one element, but found none"));
-
-        }
-    }
 
     private DatabaseFixtures.TestAccount addGatewayAccountAndCredential(
             String paymentProvider,
