+++ src/test/java/uk/gov/pay/connector/gatewayaccountcredentials/resource/GatewayAccountCredentialsResourceWorldpay3dsFlexIT.java
@@ -29,9 +29,12 @@ import static com.github.tomakehurst.wiremock.client.WireMock.serverError;
 import static io.restassured.RestAssured.given;
 import static io.restassured.http.ContentType.JSON;
 import static java.lang.String.format;
+import static javax.ws.rs.core.Response.Status.OK;
 import static org.apache.commons.lang3.RandomUtils.nextLong;
 import static org.hamcrest.CoreMatchers.is;
 import static org.hamcrest.MatcherAssert.assertThat;
+import static org.hamcrest.Matchers.hasEntry;
+import static org.junit.jupiter.api.Assertions.assertThrows;
 import static uk.gov.pay.connector.gatewayaccount.model.GatewayAccount.CREDENTIALS_MERCHANT_CODE;
 import static uk.gov.pay.connector.gatewayaccount.model.GatewayAccount.CREDENTIALS_PASSWORD;
 import static uk.gov.pay.connector.gatewayaccount.model.GatewayAccount.CREDENTIALS_USERNAME;
@@ -141,7 +144,7 @@ public class GatewayAccountCredentialsResourceWorldpay3dsFlexIT {
         }
 
         @Test
-        void forLiveAccount_shouldSetGatewayAccountCredentialsStateTo_andReturn200() throws JsonProcessingException {
+        void forLiveAccount_shouldSetGatewayAccountCredentialsStateToActive_andReturn200() throws JsonProcessingException {
             DatabaseFixtures.TestAccount testAccount = addGatewayAccountAndCredential("worldpay", CREATED, LIVE);
             String payload = objectMapper.writeValueAsString(Map.of(
                     "issuer", VALID_ISSUER,
@@ -193,6 +196,156 @@ public class GatewayAccountCredentialsResourceWorldpay3dsFlexIT {
             assertThat(result.get("issuer"), is(newIssuer));
             assertThat(result.get("organisational_unit_id"), is(newOrgUnitId));
             assertThat(result.get("jwt_mac_key"), is(updatedJwtMacKey));
+        }
+    }
+    
+    @Nested
+    class CreateOrUpdateWorldpay3DSCredentials_byServiceIdAndAccountType {
+        private static final String VALID_ISSUER = "53f0917f101a4428b69d5fb0"; // pragma: allowlist secret`
+        private static final String VALID_ORG_UNIT_ID = "57992a087a0c4849895ab8a2"; // pragma: allowlist secret`
+        private static final String VALID_JWT_MAC_KEY = "4cabd5d2-0133-4e82-b0e5-2024dbeddaa9"; // pragma: allowlist secret`
+        private final Map<String, String> valid3dsFlexCredentialsPayload = Map.of(
+                "issuer", VALID_ISSUER,
+                "organisational_unit_id", VALID_ORG_UNIT_ID,
+                "jwt_mac_key", VALID_JWT_MAC_KEY
+        );
+        @Test
+        void forNoExistingCredentials_shouldSet3dsFlexCredentials_andReturn200() {
+            String serviceId = "a-valid-service-id";
+            String gatewayAccountId = app.givenSetup()
+                    .body(toJson(Map.of(
+                            "payment_provider", "worldpay",
+                            "service_id", serviceId,
+                            "service_name", "a-test-service",
+                            "type", "test"
+                    )))
+                    .post("/v1/api/accounts")
+                    .then().extract().path("gateway_account_id");
+
+            app.givenSetup()
+                    .body(toJson(valid3dsFlexCredentialsPayload))
+                    .put(format("/v1/api/service/%s/%s/3ds-flex-credentials", serviceId, TEST))
+                    .then()
+                    .statusCode(200);
+
+            Map<String, Object> updatedGatewayAccountCredentials = app.getDatabaseTestHelper().getWorldpay3dsFlexCredentials(Long.valueOf(gatewayAccountId));
+            assertThat(updatedGatewayAccountCredentials, hasEntry("issuer", VALID_ISSUER));
+            assertThat(updatedGatewayAccountCredentials, hasEntry("jwt_mac_key", VALID_JWT_MAC_KEY));
+            assertThat(updatedGatewayAccountCredentials, hasEntry("organisational_unit_id", VALID_ORG_UNIT_ID));
+        }
+        
+        @Test
+        void shouldReturn404_forNonExistentGatewayAccount() {
+            String serviceId = "a-service-id-with-no-gateway-account";
+
+            app.givenSetup()
+                    .body(toJson(valid3dsFlexCredentialsPayload))
+                    .put(format("/v1/api/service/%s/%s/3ds-flex-credentials", serviceId, TEST))
+                    .then()
+                    .statusCode(404)
+                    .body("message[0]", is(format("Gateway account not found for service ID [%s] and account type [%s]", serviceId, TEST)));
+        }
+
+        @Test
+        void shouldReturn404_andNotUpdateCredentials_forNonWorldpayGatewayAccount() {
+            String gatewayAccountId = app.givenSetup()
+                    .body(toJson(Map.of(
+                            "payment_provider", "stripe",
+                            "service_id", "a-valid-service-id",
+                            "service_name", "a-test-service",
+                            "type", "test"
+                    )))
+                    .post("/v1/api/accounts")
+                    .then().extract().path("gateway_account_id");
+
+            app.givenSetup()
+                    .body(toJson(valid3dsFlexCredentialsPayload))
+                    .put("/v1/api/service/a-valid-service-id/test/3ds-flex-credentials")
+                    .then()
+                    .statusCode(404)
+                    .body("message[0]", is("Not a Worldpay gateway account"));
+
+            Exception exception = assertThrows(IllegalStateException.class, () ->
+                    app.getDatabaseTestHelper().getWorldpay3dsFlexCredentials(Long.valueOf(gatewayAccountId))
+            );
+            assertThat(exception.getMessage(), is("Expected at least one element, but found none"));
+        }
+
+        @Test
+        void forLiveAccount_shouldSetGatewayAccountCredentialsStateToActive_andReturn200() throws JsonProcessingException {
+            String gatewayAccountId = app.givenSetup()
+                    .body(toJson(Map.of(
+                            "payment_provider", "worldpay",
+                            "service_id", "a-valid-service-id",
+                            "service_name", "a-test-service",
+                            "type", "live"
+                    )))
+                    .post("/v1/api/accounts")
+                    .then().extract().path("gateway_account_id");
+            
+            app.givenSetup()
+                    .body(objectMapper.writeValueAsString(Map.of("payment_provider", "worldpay", "credentials", Map.of(
+                            ONE_OFF_CUSTOMER_INITIATED, toJson(Map.of(
+                                CREDENTIALS_MERCHANT_CODE, "a-merchant-code",
+                                CREDENTIALS_USERNAME, "a-username",
+                                CREDENTIALS_PASSWORD, "a-password"))))))
+                    .post("/v1/api/service/a-valid-service-id/live/credentials")
+//                    .post(format("/v1/api/accounts/%s/credentials", gatewayAccountId))
+//                    .getBody().asString();
+                    .then()
+                    .statusCode(200);
+            
+            app.givenSetup()
+                    .body(toJson(valid3dsFlexCredentialsPayload))
+                    .put("/v1/api/service/a-valid-service-id/live/3ds-flex-credentials")
+                    .then()
+                    .statusCode(200);
+
+            Map<String, Object> updatedGatewayAccountCredentials = app.getDatabaseTestHelper().getWorldpay3dsFlexCredentials(Long.valueOf(gatewayAccountId));
+            assertThat(updatedGatewayAccountCredentials, hasEntry("issuer", VALID_ISSUER));
+            assertThat(updatedGatewayAccountCredentials, hasEntry("jwt_mac_key", VALID_JWT_MAC_KEY));
+            assertThat(updatedGatewayAccountCredentials, hasEntry("organisational_unit_id", VALID_ORG_UNIT_ID));
+
+            List<Map<String, Object>> gatewayAccountCredentials = app.getDatabaseTestHelper().getGatewayAccountCredentialsForAccount(Long.valueOf(gatewayAccountId));
+            assertThat(gatewayAccountCredentials.get(0).get("state"), is("ACTIVE"));
+        }
+
+        @Test
+        void forExistingCredentials_shouldOverrideSetWorldpay3dsCredentials() {
+            String gatewayAccountId = app.givenSetup()
+                    .body(toJson(Map.of(
+                            "payment_provider", "worldpay",
+                            "service_id", "a-valid-service-id",
+                            "service_name", "a-test-service",
+                            "type", "test"
+                    )))
+                    .post("/v1/api/accounts")
+                    .then().extract().path("gateway_account_id");
+            
+            app.givenSetup()
+                    .body(toJson(valid3dsFlexCredentialsPayload))
+                    .put("/v1/api/service/a-valid-service-id/test/3ds-flex-credentials")
+                    .then()
+                    .statusCode(200);
+
+            String newIssuer = "43f0917f101a4428b69d5fb9"; // pragma: allowlist secret`
+            String newOrgUnitId = "44992a087a0c4849895cc9a3"; // pragma: allowlist secret`
+            String updatedJwtMacKey = "512ee2a9-4a3e-46d4-86df-8e2ac3d6a6a8"; // pragma: allowlist secret`
+            String payload = toJson(Map.of(
+                    "issuer", newIssuer,
+                    "organisational_unit_id", newOrgUnitId,
+                    "jwt_mac_key", updatedJwtMacKey
+            ));
+            app.givenSetup()
+                    .body(payload)
+                    .put("/v1/api/service/a-valid-service-id/test/3ds-flex-credentials")
+                    .then()
+                    .statusCode(200);
+
+            Map<String, Object> updatedGatewayAccountCredentials = app.getDatabaseTestHelper().getWorldpay3dsFlexCredentials(Long.valueOf(gatewayAccountId));
+            assertThat(updatedGatewayAccountCredentials, hasEntry("issuer", newIssuer));
+            assertThat(updatedGatewayAccountCredentials, hasEntry("jwt_mac_key", updatedJwtMacKey));
+            assertThat(updatedGatewayAccountCredentials, hasEntry("organisational_unit_id", newOrgUnitId));
         }
     }
